+++ R/profanity.R
@@ -196,7 +196,18 @@ profanity.get_sentences_data_frame <- function(text.var, profanity_list = unique
 #' @export
 plot.profanity <- function(x, transformation.function = syuzhet::get_dct_transform, ...){
 
-    m <- transformation.function(stats::na.omit(x[["profanity"]]), ...)
+
+    x <- stats::na.omit(x[["profanity"]])
+    
+    if (length(x) < 3) stop('Output contains less than 3 observations.  Cannot plot n < 3', call. = FALSE)
+    
+    if (length(x) < 100 && isTRUE(all.equal(syuzhet::get_dct_transform, transformation.function))) {
+        x <- stats::approx(x = seq_along(x), y = x, n = 100)[['y']]
+    }    
+      
+        
+    
+    m <- transformation.function(x, ...)
 
     dat <- data.frame(
         Emotional_Valence = m,